requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
django-import-export==4.3.9
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
phonenumberslite==9.0.12
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tablib==3.8.0
twilio==9.7.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
